Don't allow null dates on SubscriptionPlan (#30)
The migration sets a default start/purchase date for any subscription
plans to today and expiration dates to a year from today. This doesn't
matter in stage/prod as we have no data there, but it prevents devs
from getting DB errors locally.
Initially we wanted to allow Null dates so that ECS could edit the
fields after creating a subscription. However, we'd like to make these
fields required initially to fix errors we're seeing and to better
align with our financial reporting considerations.