Clean up email code
After changing a number of things in license-manager's assume role
policy (that will be documented in confluence), we no longer need the
SESEmailClient. The connection to SES seems to be automatically
handled by the assume role policy which connects to SES.
The only change is that we don't close the connection, and we may need
to update our email code to send to a maximum of 50 recipients at a
time.
* Also removes boto3 as a direct dependency and upgrades some packages