Allow assigning to previously revoked licenses (#61)
This is done to both keep our uniqueness constraint of one license per user
per subscription plan, and to keep the record of users that have had
licenses revoked while also allowing those users to have licenses
assigned to them again.
After a license has been revoked, the admin can specify that user in
an assignment again. On the backend, the license will be changed back
to unassigned and have all of its data cleared behind the scene, before
being assigned to the user specified.
This PR also adds to the logic that checks whether there are sufficient
licenses to assign to, as we want to consider revoked licenses in the
pool of available licenses if they were previously assigned to a user
who is now being reassigned.