Add license-subisdy api endpoint for learners
This endpoint allows a learner to lookup the subsidy data for a given
course. The endpoint looks at the given enterprise customer, gets that
customer's active subscription plan, and checks to see if a user has an
activated license for that subscription. If so, the endpoint checks to
verify that the course is in the given subscription's enterprise
catalog. If the course is in the catalog, the user has access to that
content, so the endpoint returns metadata relevant to the subsidy the
license provides.

ENT-2752