Add Salesforce, Netsuite fields to Subscriptions
Adds the salesforce_opportunity_id field to track the business
transaction of a customer purchasing a subscription. Adds the
netsuite_product_id field to track the "product/content catalog" sold
to the customer. These fields will need to be populated by an admin
from ECS using the Django Admin site at the time the subscription is
created.
The migration adds one-time default values for the salesforce and
netsuite fields to handle the existing subscriptions in our database.
As we don't have any real transactions or data yet for stage & prod
license-manager, this is just dummy data.

ENT-3090