Update api utils to 400 when keys missing from JWT
As we need the views that rely on the user's info from the JWT to also
have session authentication (in order for calls from the learner-portal
to work), there is a chance that calls from the browser may not have a
JWT associated with them. This shouldn't happen frequently (if at all),
as these views are not meant to be used from the browser, but this
update ensures that an error is raised and information is returned in
the case that it does occur.