feat: make the Braze event tracking of license assignment asynchronous.
Prior to this change, there were two synchronous requests made to the Braze tracking endpoints
for each learner being assigned in the request.  This change creates a new celery
task that wraps the ``track_license_changes()`` function.  The new task
is now called from the license assignment API action, so that requests to assign
a batch of licenses do not take too long (or time out).
Also forces ``make test`` to run pytest with the ``test`` django settings module.
ENT-5234