fix: improve the validate_num_catalog_queries mgmt command, so that it is less brittle.
The expected number of catalog queries should be determined from the distinct
number of products, not the distinct number of netsuite ids, because
many products have no netsuite id.
Also improves readability and usefulness of error message in case
the expected number of catalog queries differs from the actual number.