fix: attempt to fix bulk enrollment results link generation
Context
=======

The results log of a bulk enrollment job are stored in S3 (in addition
to being sent to celery), and we provide an API endpoint to generate a
signed download link which S3 uses to authorize temporary access to the
results object. This not only allows us to build automation to consume
job results, but currently is the ONLY way for engineers to view job
results for celery tasks that age out of Flower.

Problem
=======

Currently, it is not possible to view bulk-license-enrollment
results uploaded to S3. Attempts to generate a download link via the
following GET endpoint result in a broken URL which contains a signature
that does not match the request.

GET /api/v1/bulk-license-enrollment/?enterprise_customer_uuid=<customer-uuid>&bulk_enrollment_job_uuid=<job-uuid>

Returns:
{
    "job_id": "<job-uuid>",
    "download_url": "<Broken download URL which returns SignatureDoesNotMatch>"
}

Theory
======

This PR attempts to fix the broken signature in two ways:

1. Fix a regression in generation of the filename by adding a missing
   f-string prefix. This alone may fix the broken signature because
   the presence of certain special characters (such as curly braces) may
   cause a mismatch between the signed payload in the request vs. what
   S3 expects.

2. Add double quotes to a component of the response-content-disposition
   HTTP header.  This alone may also fix the broken signature by forcing
   certain special characters to be treated as such.  It is also more
   compliant with RFC-2616.

I don't know for sure if either will fix the issue, but they both seem
net-better.