fix: don't blank lms_user_id when clearing PII from licenses
The lms_user_id was assumed to be PII, but it actually is not.  Teams
performing data analysis have voiced a desire to have us stop clearing
lms_user_id as it is the most reliable join record to link retired
licenses to retired LMS users.

It's also just good practice to not clear records when they don't need
to be cleared.

ENT-9962